refactor: refactoring the prisma service
